Modularize Slide and Mix section
